+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+# 2.10.1
+  * Add exponential backoff delay before reconnecting resumed bulk export downloads to prevent hitting Marketo's concurrent access limit (error 615) [#118](https://github.com/singer-io/tap-marketo/pull/118)
+  * Handle Marketo error 615 (concurrent access limit) as a retryable error in stream_export with exponential backoff [#118](https://github.com/singer-io/tap-marketo/pull/118)
+
 # 2.10.0
   * Python Upgrade, Dependency Upgrade, Added unit tests and mocked integration tests. [#119](https://github.com/singer-io/tap-marketo/pull/119)